+++ demo/CalendarDemo.jsx
@@ -179,6 +179,7 @@ class Demo extends React.Component {
             showSecond
             showHour
             showTime
+            // renderSidebar={this.sidebarRender}
             size="small"
             timezone={8}
             defaultOpenValue={new Date(2018, 8, 1, 3, 12)}
@@ -187,17 +188,17 @@ class Demo extends React.Component {
             onSelect={this.onSelect}
           />
         </div>
-        {/* <div
-          className="kuma-form-field"
-          style={{
-            width: '400px',
-          }}
-        >
-          <p>
-            禁用
-          </p>
-          <Calendar value={value} disabled onSelect={this.onSelect} />
-        </div>
+        {/*<div*/}
+          {/*className="kuma-form-field"*/}
+          {/*style={{*/}
+            {/*width: '400px',*/}
+          {/*}}*/}
+        {/*>*/}
+          {/*<p>*/}
+            {/*禁用*/}
+          {/*</p>*/}
+          {/*<Calendar value={value} disabled onSelect={this.onSelect} />*/}
+        {/*</div>*/}
         <div
           className="kuma-form-field"
           style={{
@@ -211,6 +212,8 @@ class Demo extends React.Component {
             size="middle"
             value={value}
             disabledDate={disabledDate}
+            allowedMonthRange={[ { start: '2018-11', end: '2018-12' }, { end: '2016-12' } ]}
+            defaultOpenValue={new Date(2016, 8, 1, 3, 12)}
             onSelect={this.onSelect}
             showDateInput
           />
@@ -231,45 +234,45 @@ class Demo extends React.Component {
             disabledDate={(current) => { }}
           />
         </div>
-        <div
-          className="kuma-form-field"
-          style={{
-            width: '400px',
-          }}
-        >
-          <p>
-            显示日期和日程
-          </p>
-          <p>
-            Calendar 通过开放 contentRender 参数来完成日期渲染上的定制，并提供了一个默认的渲染函数 Calendar.util.generateContentRender(code) 来完成通用定制。
-          </p>
-          <Calendar
-            value={value}
-            onSelect={this.onSelect}
-            contentRender={Calendar.util.generateContentRender(
-              {
-                '2016-01-07': 'leave',
-                '2016-01-09': 'work',
-                '2016-01-08': 'schedule',
-              },
-              'en',
-            )}
-          />
-        </div>
-        <div
-          className="kuma-form-field"
-          style={{
-            width: '400px',
-          }}
-        >
-          <p>
-            直接渲染面板
-          </p>
-          <RcCalendar {...panelOptions} className="panel-demo" />
-        </div>
-        <Button onClick={me.handleClick}>
-          changeTime
-        </Button> */}
+        {/*<div*/}
+          {/*className="kuma-form-field"*/}
+          {/*style={{*/}
+            {/*width: '400px',*/}
+          {/*}}*/}
+        {/*>*/}
+          {/*<p>*/}
+            {/*显示日期和日程*/}
+          {/*</p>*/}
+          {/*<p>*/}
+            {/*Calendar 通过开放 contentRender 参数来完成日期渲染上的定制，并提供了一个默认的渲染函数 Calendar.util.generateContentRender(code) 来完成通用定制。*/}
+          {/*</p>*/}
+          {/*<Calendar*/}
+            {/*value={value}*/}
+            {/*onSelect={this.onSelect}*/}
+            {/*contentRender={Calendar.util.generateContentRender(*/}
+              {/*{*/}
+                {/*'2016-01-07': 'leave',*/}
+                {/*'2016-01-09': 'work',*/}
+                {/*'2016-01-08': 'schedule',*/}
+              {/*},*/}
+              {/*'en',*/}
+            {/*)}*/}
+          {/*/>*/}
+        {/*</div>*/}
+        {/*<div*/}
+          {/*className="kuma-form-field"*/}
+          {/*style={{*/}
+            {/*width: '400px',*/}
+          {/*}}*/}
+        {/*>*/}
+          {/*<p>*/}
+            {/*直接渲染面板*/}
+          {/*</p>*/}
+          {/*<RcCalendar {...panelOptions} className="panel-demo" />*/}
+        {/*</div>*/}
+        {/*<Button onClick={me.handleClick}>*/}
+          {/*changeTime*/}
+        {/*</Button>*/}
         <div
           className="kuma-form-field"
           style={{